pyproject.toml
[project]
name = "cryto-agency"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["google-adk>=1.16.0", "requests>=2.31.0"]
